backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
snowflake-connector-python==3.6.0
python-dotenv==1.0.0
pandas==2.1.4
pydantic==2.5.3
pydantic-settings==2.1.0

